+++ client/src/screens/hub/HubMainScreen/components/ControlsSidebar/ControlsSidebar.js
@@ -22,12 +22,15 @@ function ControlsSidebar() {
     runs,
     chart,
     contextFilter,
+    colorPalette,
   } = HubMainScreenModel.useHubMainScreenState([
     HubMainScreenModel.events.SET_RUNS_STATE,
     HubMainScreenModel.events.SET_CHART_SETTINGS_STATE,
     HubMainScreenModel.events.SET_CHART_POINTS_COUNT,
     HubMainScreenModel.events.SET_CONTEXT_FILTER,
     HubMainScreenModel.events.SET_SEED,
+    HubMainScreenModel.events.TOGGLE_PERSISTENCE,
+    HubMainScreenModel.events.SET_COLOR_PALETTE,
   ]);
 
   let {
@@ -41,6 +44,7 @@ function ControlsSidebar() {
     groupByChart,
     aggregated,
     seed,
+    persist,
   } = contextFilter;
 
   return (
@@ -60,8 +64,17 @@ function ControlsSidebar() {
         </ContentLoader>
       ) : (
         <div className='ControlsSidebar__items'>
-          <GroupByColor groupByColor={groupByColor} seed={seed.color} />
-          <GroupByStyle groupByStyle={groupByStyle} seed={seed.style} />
+          <GroupByColor
+            groupByColor={groupByColor}
+            seed={seed.color}
+            persist={persist.color}
+            colorPalette={colorPalette}
+          />
+          <GroupByStyle
+            groupByStyle={groupByStyle}
+            seed={seed.style}
+            persist={persist.style}
+          />
           <GroupByChart groupByChart={groupByChart} />
           {isExploreMetricsModeEnabled() && (
             <>