+++ client/src/screens/hub/HubMainScreen/components/ControlsSidebar/components/GroupByColor/GroupByColor.js
@@ -1,19 +1,28 @@
+import './GroupByColor.less';
+
 import React, { useState, useRef, useEffect } from 'react';
 import UI from '../../../../../../../ui';
 import PropTypes from 'prop-types';
 import { classNames } from '../../../../../../../utils';
 import { getGroupingOptions } from '../../helpers';
 import { HubMainScreenModel } from '../../../../models/HubMainScreenModel';
+import { COLORS } from '../../../../../../../constants/colors';
 
 function GroupByColor(props) {
   let [opened, setOpened] = useState(false);
+  let [advanceOpened, setAdvanceOpened] = useState(false);
 
-  const { groupByColor, seed } = props;
+  const { groupByColor, seed, persist, colorPalette } = props;
 
   let popupRef = useRef();
   let dropdownRef = useRef();
 
-  let { setContextFilter, setSeed } = HubMainScreenModel.emitters;
+  let {
+    setContextFilter,
+    setSeed,
+    togglePersistence,
+    setColorPalette,
+  } = HubMainScreenModel.emitters;
 
   let {
     getAllParamsPaths,
@@ -106,15 +115,114 @@ function GroupByColor(props) {
               isOpen
               multi
             />
-            <UI.Line />
-            <div className='ControlsSidebar__item__popup__body__action'>
-              <UI.Button
-                size='tiny'
-                disabled={groupByColor.length === 0}
-                onClick={(evt) => setSeed(seed + 1, 'color')}
+            <div className='ControlsSidebar__item__popup__body__actionContainer'>
+              <UI.Tooltip
+                tooltip={
+                  advanceOpened
+                    ? 'Hide advanced options'
+                    : 'Show advanced options'
+                }
               >
-                Shuffle groups colors
-              </UI.Button>
+                <div
+                  className={classNames({
+                    ControlsSidebar__item__popup__body__actionCollapse: true,
+                    collapsed: advanceOpened,
+                  })}
+                  onClick={() => setAdvanceOpened((opened) => !opened)}
+                >
+                  <span className='ControlsSidebar__item__popup__body__toggle__action'>
+                    <UI.Icon
+                      i={advanceOpened ? 'unfold_less' : 'unfold_more'}
+                      scale={1}
+                    />
+                  </span>
+                  <UI.Text type='grey-dark' small>
+                    Advanced options
+                  </UI.Text>
+                </div>
+              </UI.Tooltip>
+              {advanceOpened && (
+                <>
+                  <div className='ControlsSidebar__item__popup__body__action'>
+                    <UI.Text overline bold type='primary'>
+                      Colors persistence:
+                    </UI.Text>
+                    <UI.Text small spacingTop spacing>
+                      Enable persistent coloring mode so that each group always
+                      has the same color regardless of its order.
+                    </UI.Text>
+                    <div className='ControlsSidebar__item__popup__body__action__row ControlsSidebar__item__popup__body__action__row--persistence'>
+                      <div
+                        className='ControlsSidebar__item__popup__body__action__row__persistenceSwitch'
+                        onClick={() => togglePersistence('color')}
+                      >
+                        <span
+                          className={classNames({
+                            ControlsSidebar__item__popup__toggle: true,
+                            on: persist,
+                          })}
+                        >
+                          <UI.Icon
+                            i={`toggle_${persist ? 'on' : 'off'}`}
+                            scale={1.5}
+                          />
+                        </span>
+                        <UI.Text type={persist ? 'primary' : 'grey-dark'} small>
+                          {persist ? 'Enabled' : 'Disabled'}
+                        </UI.Text>
+                      </div>
+                      {persist && (
+                        <UI.Button
+                          size='tiny'
+                          disabled={groupByColor.length === 0}
+                          onClick={(evt) => setSeed(seed + 1, 'color')}
+                        >
+                          Shuffle colors
+                        </UI.Button>
+                      )}
+                    </div>
+                  </div>
+                  <div className='ControlsSidebar__item__popup__body__action'>
+                    <UI.Text overline bold type='primary'>
+                      Preferred color palette:
+                    </UI.Text>
+                    {COLORS.map((palette, paletteIndex) => (
+                      <div
+                        key={paletteIndex}
+                        className='ColorPalette'
+                        onClick={() => setColorPalette(paletteIndex)}
+                      >
+                        <UI.Radio
+                          name={paletteIndex}
+                          checked={+paletteIndex === +colorPalette}
+                        />
+                        <div
+                          className={`ColorPalette__colors ColorPalette__colors--${paletteIndex}`}
+                        >
+                          {palette.map((color) => (
+                            <span
+                              className='ColorPalette__colors__item'
+                              style={{
+                                backgroundColor: color,
+                              }}
+                              key={color}
+                            />
+                          ))}
+                        </div>
+                        <UI.Text
+                          inline
+                          small
+                          type='grey-dark'
+                          className='ColorPalette__colors__title'
+                        >
+                          {paletteIndex === 0 && ' 8 distinct colors'}
+                          {paletteIndex === 1 && ' 24 colors'}
+                        </UI.Text>
+                      </div>
+                    ))}
+                  </div>
+                </>
+              )}
             </div>
           </div>
         </div>
@@ -126,6 +234,8 @@ function GroupByColor(props) {
 GroupByColor.propTypes = {
   groupByColor: PropTypes.arrayOf(PropTypes.string),
   seed: PropTypes.number,
+  persist: PropTypes.bool,
+  colorPalette: PropTypes.number,
 };
 
 export default GroupByColor;