+++ client/src/screens/hub/HubMainScreen/components/Panel/components/PanelChart/PanelChart.js
@@ -236,9 +236,9 @@ function PanelChart(props) {
         .text(
           traceList?.grouping.chart.length > 0
             ? `#${props.index + 1} ${traceList?.grouping.chart
-              .map((key) => {
-                return (
-                  key +
+                .map((key) => {
+                  return (
+                    key +
                     '=' +
                     formatValue(
                       traceList.traces.find(
@@ -246,18 +246,18 @@ function PanelChart(props) {
                       )?.config[key],
                       false,
                     )
-                );
-              })
-              .join(', ')}`
+                  );
+                })
+                .join(', ')}`
             : '',
         )
         .append('svg:title')
         .text(
           traceList?.grouping.chart.length > 0
             ? `#${props.index + 1} ${traceList?.grouping.chart
-              .map((key) => {
-                return (
-                  key +
+                .map((key) => {
+                  return (
+                    key +
                     '=' +
                     formatValue(
                       traceList.traces.find(
@@ -265,9 +265,9 @@ function PanelChart(props) {
                       )?.config[key],
                       false,
                     )
-                );
-              })
-              .join(', ')}`
+                  );
+                })
+                .join(', ')}`
             : '',
         );
     }